feat(expedition): Integrate expedition system into game loop
APP MODE UPDATES:
- Added 3 new AppMode variants:
  - ExpeditionSetup: Size selection and risk assessment
  - ExpeditionInProgress: Animation-only battle display
  - ExpeditionSummary: Loot reveal and results

GAME MANAGER UPDATES:
- Added expedition_setup_page: Option<ExpeditionSetupPage>
- Added expedition_summary_page: Option<ExpeditionSummaryPage>
- Added expedition_data: Option<ExpeditionData>

EXPEDITION DATA:
- Tracks ongoing expedition state:
  - enemy_id, enemy_name
  - target_kills, expedition_size
  - start_time, end_time
  - ExpeditionResult (pre-calculated)

RESOURCE INTEGRATION:
- Updated get_current_page() to handle all 3 expedition modes
- ExpeditionInProgress reuses battle_page for animation display
- All GameManager::new() and from_save_data() updated

RENDER SYSTEM:
- Added expedition modes to standard page rendering flow
- Uses same rendering pipeline as Battle/BattleResult
- Page trait update/draw cycle works for all expedition pages

TYPE FIXES:
- Fixed bar_width type cast in expedition_summary.rs
- Changed bar_width / 2 to (bar_width / 2) as i32

COMPILATION STATUS:
✅ Compiles successfully with 0 errors
⚠️ Only unused import warnings remain

NEXT STEPS:
- Update map navigation to create expedition flow
- Implement expedition start logic
- Add expedition progress system
- Integrate 10-minute KO recovery

+++ tamagotchi/stdgotchi/src/ui/pages/expedition_summary.rs
@@ -179,7 +179,7 @@ impl ExpeditionSummaryPage {
         let mut progress_text = heapless::String::<32>::new();
         write!(progress_text, "{}/{}", current_exp, exp_to_level).ok();
         let text_style = MonoTextStyle::new(&FONT_6X10, Rgb888::WHITE);
-        let text_x = bar_x + (bar_width / 2) - ((progress_text.len() as i32) * 3);
+        let text_x = bar_x + ((bar_width / 2) as i32) - ((progress_text.len() as i32) * 3);
         Text::new(&progress_text, Point::new(text_x, bar_y + 13), text_style).draw(display)?;
 
         Ok(())
